# Optimize multiplication for Normed