This adds `wrapping_mul`, `saturating_mul` and `checked_mul` binary operations. However, this does not specialize them for `Fixed` and does not change [`*` for `Fixed`](https://github.com/JuliaMath/FixedPointNumbers.jl/blob/5dcaf979ac148012d7d95b018232c9a86d7416a9/src/fixed.jl#L103).

This applies the wrapping arithmetic as the default arithmetic of `*` for the abstract `FixedPoint`, but as mentioned above, the `*`for `Fixed` is not affected by the default arithmetic. Furthermore, `*` for `Normed` overrides the default arithmetic with the "checked" arithmetic for backward compatibility. (Strictly speaking, the error type is changed from `ArgumentError` to `OverflowError`.)

This replaces most of Normed's implementation of multiplication with integer operations. This improves the speed in many cases and the accuracy in some cases.

Fixes #174