@kimikage 2020-08-14
## Benchmark
```julia
x_n0f8  = collect(rand(N0f8,  1000, 1000)); y_n0f8  = collect(rand(N0f8,  1000, 1000));
x_n0f16 = collect(rand(N0f16, 1000, 1000)); y_n0f16 = collect(rand(N0f16, 1000, 1000));
x_n4f4  = collect(rand(N4f4,  1000, 1000)); y_n4f4  = collect(rand(N4f4,  1000, 1000));
x_n4f12 = collect(rand(N4f12, 1000, 1000)); y_n4f12 = collect(rand(N4f12, 1000, 1000));
z_n4f4  = N4f4.( rand(1000, 1000));
z_n4f12 = N4f12.(rand(1000, 1000));

@btime $x_n0f8 .* $y_n0f8;
@btime $x_n0f16 .* $y_n0f16;
@btime saturating_mul.($x_n4f4, $y_n4f4);
@btime saturating_mul.($x_n4f12, $y_n4f12);
@btime $x_n4f4 .* $z_n4f4;
@btime $x_n4f12 .* $z_n4f12;
```

| | Julia v1.5.0<br>before | Julia v1.5.0<br>after | Julia v1.4.2<br>before | Julia v1.4.2<br>after |
|-----------------------------------|---------:|-----------:|---------:|-----------:|
|`x_n0f8 .* y_n0f8`                 | 1.965 ms |  77.600 μs | 1.978 ms |  85.100 μs |
|`x_n0f16 .* y_n0f16`               | 2.172 ms | 165.900 μs | 2.168 ms | 166.200 μs |
|`saturating_mul.(x_n4f4, y_n4f4)`  |        - | 105.900 μs |        - | 104.100 μs |
|`saturating_mul.(x_n4f12, y_n4f12)`|        - | 304.200 μs |        - | 302.700 μs |
|`$x_n4f4 .* $z_n4f4` (checked_mul) | 2.018 ms | 794.400 μs | 1.991 ms | 794.700 μs |
|`$x_n4f12 .* $z_n4f12`(checked_mul)| 2.636 ms |   1.086 ms | 2.688 ms |   1.090 ms |