@kimikage 2020-08-24
Thank you for all your reviews.
As for the default arithmetic settings, I will ask for feedback again (in Discourse?) before the release of v0.9.